+++ src/main/java/za/ac/cput/factory/StudentFactory.java
@@ -3,7 +3,7 @@ package za.ac.cput.factory;
 import za.ac.cput.domain.Student;
 
 public class StudentFactory {
-    public Student createStudent(String firstName, String lastName,String phoneNumber, String password, String email, Integer studentNumber) {
+    public Student createStudent(String firstName, String lastName,String phoneNumber, String password, String email, Long studentNumber) {
         return Student.builder()
                 .firstName(firstName)
                 .lastName(lastName)
