+++ src/main/java/za/ac/cput/repository/TutorRepository.java
@@ -3,5 +3,5 @@ package za.ac.cput.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import za.ac.cput.domain.Tutor;
 
-public interface TutorRepository extends JpaRepository<Tutor, Integer> {
+public interface TutorRepository extends JpaRepository<Tutor, Long> {
 }
